requirements.txt
fastapi
sqlmodel
jinja2
passlib[bcrypt]
python-jose